Switched the configuration from vim script to lua
To make writing the configuration easier I switched the configuration
from vim script to lua.
This makes the config a little less verbose removing all 'nore' from the
mappings.
What is not great at the moment are the commands, which are a wild mix
of lua and vim script. Hopefully this will improve with later iterations